+++ sampledb/frontend/objects.py
@@ -286,7 +286,6 @@ def objects():
             else:
                 sorting_order_name = 'asc'
                 sorting_order = object_sorting.ascending
-
         if sorting_property_name is None:
             sorting_property_name = '_object_id'
         else:
@@ -1368,6 +1367,7 @@ def object(object_id):
         return flask.render_template(
             'objects/view/base.html',
             template_mode="view",
+            show_object_type_and_id_on_object_page_text=get_user_settings(flask_login.current_user.id)["SHOW_OBJECT_TYPE_AND_ID_ON_OBJECT_PAGE"],
             show_object_title=get_user_settings(flask_login.current_user.id)["SHOW_OBJECT_TITLE"],
             measurement_type_name=logic.action_type_translations.get_action_type_translation_for_action_type_in_language(
                 action_type_id=logic.actions.models.ActionType.MEASUREMENT,
@@ -2093,6 +2093,7 @@ def object_version(object_id, version_id):
     return flask.render_template(
         'objects/view/base.html',
         template_mode="view",
+        show_object_type_and_id_on_object_page_text=get_user_settings(flask_login.current_user.id)["SHOW_OBJECT_TYPE_AND_ID_ON_OBJECT_PAGE"],
         show_object_title=get_user_settings(flask_login.current_user.id)["SHOW_OBJECT_TITLE"],
         languages=languages,
         metadata_language=metadata_language,